refactor: ctf dto 패키지 정리
