+++ src/main/java/com/keeper/homepage/domain/ctf/api/CtfTeamController.java
@@ -1,11 +1,11 @@
 package com.keeper.homepage.domain.ctf.api;
 
 import com.keeper.homepage.domain.ctf.application.CtfTeamService;
-import com.keeper.homepage.domain.ctf.dto.request.CreateTeamRequest;
-import com.keeper.homepage.domain.ctf.dto.request.JoinTeamRequest;
-import com.keeper.homepage.domain.ctf.dto.request.UpdateTeamRequest;
-import com.keeper.homepage.domain.ctf.dto.response.CtfTeamDetailResponse;
-import com.keeper.homepage.domain.ctf.dto.response.CtfTeamResponse;
+import com.keeper.homepage.domain.ctf.dto.request.team.CreateTeamRequest;
+import com.keeper.homepage.domain.ctf.dto.request.team.JoinTeamRequest;
+import com.keeper.homepage.domain.ctf.dto.request.team.UpdateTeamRequest;
+import com.keeper.homepage.domain.ctf.dto.response.team.CtfTeamDetailResponse;
+import com.keeper.homepage.domain.ctf.dto.response.team.CtfTeamResponse;
 import com.keeper.homepage.domain.member.entity.Member;
 import com.keeper.homepage.global.config.security.annotation.LoginMember;
 import jakarta.validation.Valid;
