+++ src/main/java/com/keeper/homepage/domain/ctf/application/CtfTeamService.java
@@ -6,8 +6,8 @@ import static com.keeper.homepage.global.error.ErrorCode.CTF_TEAM_INACCESSIBLE;
 import com.keeper.homepage.domain.ctf.application.convenience.CtfContestFindService;
 import com.keeper.homepage.domain.ctf.application.convenience.CtfTeamFindService;
 import com.keeper.homepage.domain.ctf.dao.team.CtfTeamRepository;
-import com.keeper.homepage.domain.ctf.dto.response.CtfTeamDetailResponse;
-import com.keeper.homepage.domain.ctf.dto.response.CtfTeamResponse;
+import com.keeper.homepage.domain.ctf.dto.response.team.CtfTeamDetailResponse;
+import com.keeper.homepage.domain.ctf.dto.response.team.CtfTeamResponse;
 import com.keeper.homepage.domain.ctf.entity.CtfContest;
 import com.keeper.homepage.domain.ctf.entity.team.CtfTeam;
 import com.keeper.homepage.domain.member.application.convenience.MemberFindService;
