--- src/main/java/com/keeper/homepage/domain/ctf/dto/response/CtfChallengeResponse.java
@@ -1,5 +1,0 @@
-package com.keeper.homepage.domain.ctf.dto.response;
-
-public class CtfChallengeResponse {
-
-}
