+++ src/main/java/com/keeper/homepage/domain/ctf/dto/response/challenge/CtfChallengeResponse.java
@@ -1,0 +1,5 @@
+package com.keeper.homepage.domain.ctf.dto.response.challenge;
+
+public class CtfChallengeResponse {
+
+}
